+++ packages/livekit-rtc/tsup.config.ts
@@ -4,5 +4,8 @@ import defaults from '../../tsup.config';
 
 export default defineConfig({
   ...defaults,
+  // Emit a single CJS implementation only with a thin ESM wrapper
+  // to avoid dual-package hazard
+  format: ['cjs'],
   external: [/\.\/.*\.cjs/, /\.\/.*.node/],
 });